src/BackEnd/requirements_backend.txt
fastapi
sqlalchemy
pydantic
requests
psycopg2
python-dotenv